Moved the UserPageFrame and UserChangeBudgetFrame to the correct package so that we adhere to clean architecture and the new past orders frame was also created so that the user will be able to view their past order
