+++ src/main/java/ui/PastOrdersFrame.java
@@ -1,0 +1,4 @@
+package ui;
+
+public class PastOrdersFrame {
+}
